backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
cryptography==43.0.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.5
greenlet==3.1.1
groq==0.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.10.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
